--- src/Arius.Core/Shared/HashCache/HashCacheEntry.cs
@@ -1,18 +1,0 @@
-using Arius.Core.Shared.FileSystem;
-using Arius.Core.Shared.Hashes;
-
-namespace Arius.Core.Shared.HashCache;
-
-/// <summary>One persisted hashcache row: the cheap signals + sparse fingerprint + cached content hash for a path.</summary>
-internal readonly record struct HashCacheEntry(
-    RelativePath Path,
-    long         Size,
-    long         MtimeTicks,
-    long?        CtimeTicks,
-    string?      Inode,
-    string?      Dev,
-    int          SignalSet,
-    byte[]       SparseFingerprint,
-    int          FpAlgo,
-    ContentHash  ContentHash,
-    long         LastVerifiedTicks);
