+++ src/Arius.Core/Shared/HashCache/HashCacheService.cs
@@ -1,7 +1,12 @@
-using Arius.Core.Shared.FileSystem;
-using Arius.Core.Shared.Hashes;
-
 namespace Arius.Core.Shared.HashCache;
+
+[SharedWithinAssembly]
+internal interface IHashCacheService
+{
+    FastHashResult TryReuse(RelativeFileSystem fs, RelativePath path, long liveSize, long now);
+    void Record(RelativePath path, long size, FileChangeSignals? signals, byte[] sparseFingerprint, ContentHash hash, long now);
+}
+
 
 /// <summary>
 /// Per-repository fast-hash facade: the verdict ladder over <see cref="HashCacheLocalStore"/>.
