--- src/Arius.Core/Shared/HashCache/IHashCacheService.cs
@@ -1,11 +1,0 @@
-using Arius.Core.Shared.FileSystem;
-using Arius.Core.Shared.Hashes;
-
-namespace Arius.Core.Shared.HashCache;
-
-[SharedWithinAssembly]
-internal interface IHashCacheService
-{
-    FastHashResult TryReuse(RelativeFileSystem fs, RelativePath path, long liveSize, long now);
-    void Record(RelativePath path, long size, FileChangeSignals? signals, byte[] sparseFingerprint, ContentHash hash, long now);
-}
