--- src/Arius.Core/Shared/HashCache/SignalSets.cs
@@ -1,10 +1,0 @@
-namespace Arius.Core.Shared.HashCache;
-
-/// <summary>Provenance tag stored on a hashcache row so signals are only compared within the same source.</summary>
-[SharedWithinAssembly]
-internal static class SignalSets
-{
-    public const int None    = 0;
-    public const int Posix   = 1;
-    public const int Windows = 2;
-}
